requirements.txt
google-genai
streamlit
pyperclip 
streamlit-ace